Renamed the project and updated the namespaces again.
