+++ TestRail.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.27004.2009
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TestRailStandard", "TestRailStandard\TestRailStandard.csproj", "{2FDADC24-EE2C-4326-94AC-148901F25673}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TestRail", "TestRail\TestRail.csproj", "{2FDADC24-EE2C-4326-94AC-148901F25673}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
