+++ TestRail/Types/Case.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using Newtonsoft.Json.Linq;
 
-namespace TestRailStandard.Types
+namespace TestRail.Types
 {
     /// <summary>stores information about a case</summary>
     public class Case : BaseTestRailType
