+++ TestRail/Types/ConfigurationGroup.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using Newtonsoft.Json.Linq;
 
-namespace TestRailStandard.Types
+namespace TestRail.Types
 {
     public class ConfigurationGroup : BaseTestRailType
     {
