+++ TestRail/Types/Milestone.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using Newtonsoft.Json.Linq;
 
-namespace TestRailStandard.Types
+namespace TestRail.Types
 {
     /// <summary>stores information about a milestone</summary>
     public class Milestone : BaseTestRailType
