+++ TestRail/Types/Result.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using Newtonsoft.Json.Linq;
 
-namespace TestRailStandard.Types
+namespace TestRail.Types
 {
     /// <summary>stores information about the result of a test</summary>
     public class Result : BaseTestRailType
